Bug fix in error return of CV read
Fixed <R> command that reads a CV so that it now returns (-1) upon
verification error instead of (255).  Problem was return value declared
as byte, not integer in this specific part of the code (in which case a
(-1 )is incorrectly interpreted as a 255).

+++ DCCpp_Uno/PacketRegister.cpp
@@ -208,7 +208,7 @@ void RegisterList::writeTextPacket(char *s) volatile{
 
 void RegisterList::readCV(char *s) volatile{
   byte bRead[4];
-  byte bValue;
+  int bValue;
   unsigned long c;
   int cv, callBack, callBackSub;
